+++ lib/karafka/pro/processing/strategies/dlq/lrj_mom.rb
@@ -37,7 +37,7 @@ module Karafka
 
                   return if coordinator.manual_pause?
 
-                  seek(last_group_message.offset + 1) unless revoked?
+                  seek(last_group_message.offset + 1, false) unless revoked? || coordinator.manual_seek?
 
                   resume
                 elsif coordinator.pause_tracker.attempt <= topic.dead_letter_queue.max_retries